Fix crash related to invalid RowSlot or RowSlotChunk pointers
Details:
Fix two issues where a RowSlot* or RowSlotChunk* could point
to invalid addresses following a CursorWindow resize which moves
the entire allocation block, invalidating the previous address.
These scenarios could occur when a query required expanding the default
CursorWindow size and the allocation occured on an edge boundary.
This adjustment also allows us to remove the CURSOR_SIZE_EXTRA which
caused the test `testManyRowsLong` to identify the RowSlotChunk issue.
The test `shouldNotCauseRowSlotAllocationCrash` verified the fix
for the RowSlot error.